backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
alembic
python-dotenv
requests
feedparser
python-dateutil
httpx
yfinance
numpy
pandas
psycopg2-binary
apscheduler
